# perf(compactor): optimize allocations: use buffer for priorities