--- a/db.go
+++ b/db.go
@@ -1617,7 +1617,7 @@ func (db *DB) Flatten(workers int) error {
 			}
 		}
 		if len(levels) <= 1 {
-			prios := db.lc.pickCompactLevels()
+			prios := db.lc.pickCompactLevels(nil)
 			if len(prios) == 0 || prios[0].score <= 1.0 {
 				db.opt.Infof("All tables consolidated into one level. Flattening done.\n")
 				return nil