@harshil-goel 2023-09-13
Thanks a lot for the diff. From what I understand you are trying to reuse the memory of the levels object, not their values correct? Do you a SyncPool might be able to do the trick better?