@deff7 2023-09-13
> Thanks a lot for the diff. From what I understand you are trying to reuse the memory of the levels object, not their values correct? Do you a SyncPool might be able to do the trick better?

Hello! I'm trying to reuse the whole slice of values. Fortunately, as far as I know, all the values from this slice are used after `runOnce` completes and no single value escapes the routine. 

I prefer using simple buffer reusing technique over sync.Pool, because it more straightforward and controllable, also we don't have any concurrency here